fix relative imports fix deps add flake8 setup
