+++ python-modules/cis_publisher/cis_publisher/__init__.py
@@ -2,25 +2,8 @@
 
 """Top-level package for cis_publisher."""
 
-__author__ = """Andrew Krug"""
-__email__ = 'akrug@mozilla.com'
-__version__ = '0.0.1'
-
-import os
-
-from everett.manager import ConfigManager
-from everett.manager import ConfigIniEnv
-from everett.manager import ConfigOSEnv
+from cis_publisher import operation
+from cis_publisher import common
 
 
-def get_config():
-    return ConfigManager(
-        [
-            ConfigIniEnv([
-                os.environ.get('CIS_CONFIG_INI'),
-                '~/.mozilla-cis.ini',
-                '/etc/mozilla-cis.ini'
-            ]),
-            ConfigOSEnv()
-        ]
-    )
+__all__ = [common, operation]
