+++ python-modules/cis_publisher/setup.cfg
@@ -1,0 +1,4 @@
+[flake8]
+exclude = docs
+max-line-length = 120
+ignore=E266
